Stairs with EMS
I don't like this solution of including different variants of SaveUnitExpanded based on whether Stairs and Strmag are included, but I'm not sure there's a better way to get it done. Stairs needs to save the stair byte in the character struct, and Strmag needs to save the magic stat.