Healtiles
This isn't too big, but this adds a command to add healing tiles to the map via trap data with a displayed map sprite indicating them. (I hijacked DVs quite a bit for this.) Define an X coordinate, a Y coordinate, and a percent of HP to heal. You can also optionally define an event ID to turn them on and off (which also affects the map sprite appearing).