Skill definitions reorg
Adds skill descriptions in next to definitions for each skill, as well as its author. Author is unknown for a good number of skills, if someone with knowledge of who made which skills would like to go through and fill them in / correct any errors that exist currently that would be much appreciated.