Add AI heatmap optimization
By moving a call to GetUnitPower out of a loop and instead storing its value to a high register, we can reduce the time needed to build the AI heatmap by 50% or more. In realtime this means we can shave anywhere from 3 to 6 seconds off each AI heatmap call, which adds up very quickly on large and/or dense maps. By removing some compiler garbage and refactoring the vanilla function, there is just enough room to free up r11 to store the result of GetUnitPower and restore it at the end of the function.