Avoid overlapping digits.
We heighten digits that might overlap with another displayed damage number.

At the moment, standard height of digits mean they can overlap with the procced-skill bar. Could raise, lower standard height of digits (in which case they won't align with MISS! and NODAMAGE!). Of course digits will be raised 16 px if they overlap, so keep that in mind.

It's possible for MISS! and NODAMAGE! to appear whilst digits are still displayed. Need to prevent this (kill moji proc in some battle anim command?)